Add testing toolchain for signal forwarding feature.
The toolchain adds a Node app (written in TypeScript) that calls inject
which then invokes the signals.sh script. The toolchain app then sends
each of SIGHUP, SIGINT, and SIGTERM to separate invocations of inject.

Unless all calls succeed the app will exit with an error. This should be
sufficient for CI testing.