Add signal support (#1)
* Update deps.

* Update repo references.

* Fix missing HelpName in main().

* Add script for testing signal forwarding.

This script is meant to act as the client, that is, the application that is
wrapped by an invocation of inject.

* Add signal forwarding support to inject.

* Add testing toolchain for signal forwarding feature.

The toolchain adds a Node app (written in TypeScript) that calls inject
which then invokes the signals.sh script. The toolchain app then sends
each of SIGHUP, SIGINT, and SIGTERM to separate invocations of inject.

Unless all calls succeed the app will exit with an error. This should be
sufficient for CI testing.

* Update Github Actions test workflow to test signal forwarding.

* Update Makefile setup-lint command to use `install` instead of `get`.

* Remove TODO in main.go.

* Fix typo in package.json.

* Update README and license.